modify : change config method
1. add mode attribute in config file
2. move config read file from DB package to tools package
3. modify the DB setup in main.go to let it distinguish when to use dummy DB or normal DB
