+++ backend/config/db copy.yaml
@@ -3,3 +3,4 @@ password: Kar98Pro?
 host: mysql
 port: 3306
 database: test
+mode : dev # dev or test
